@@ -4,11 +4,12 @@
   "private": true,
   "scripts": {
     "dev": "next dev -p 3002",
-    "build": "next build",
+    "build": "IS_BUILD=true next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
+    "aos": "^2.3.4",
     "axios": "^1.7.7",
     "next": "15.0.2",
     "react": "19.0.0-rc-02c0e824-20241028",
@@ -16,11 +17,12 @@
     "telegram": "^2.26.6"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
+    "@types/aos": "^3.0.7",
+    "@types/node": "22.9.0",
+    "@types/react": "18.3.12",
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "15.0.2",
-    "typescript": "^5"
+    "typescript": "5.6.3"
   }
 }
