@@ -1,7 +1,8 @@
 // pages/_app.tsx
+import { AppProps } from "next/app";
 import Layout from "./components/Layout";
 
-function MyApp({ Component, pageProps }) {
+function MyApp({ Component, pageProps }: AppProps) {
   return (
     <Layout>
       <Component {...pageProps} />
